[clang-format] Treat AttributeMacros more like attribute macros in a few cases.
I noticed that clang-format was inserting some strange indentation whenever I used custom "attribute-like macros"
(things like `FOO_EXTERN` to wrap attribute-visible-default, or macros with parentheses like `NS_SWIFT_NAME(...)`).

There are two parts to this fix:
 - tokenize the paren after an `AttributeMacro` as a `TT_AttributeParen`
 - treat a `AttributeMacro`-without-paren the same as one with a paren (eg, the `FOO_EXTERN` case)

There are new tests cases to differentiate macros that are (or are-not) `AttributeMacro` and also handle when
`ColumnLimit` is set to infinite ("0") or a finite value, as part of this patch is in `ContinuationIndenter`.

Differential Revision: https://reviews.llvm.org/D145262